feat(metadata): schema API client
Implements a GetSchema() metadata call that returns the set of columns
and their data types for a specific table within a namespace.

Redefines the column data types defined in the proto spec, decoupling
callers from the generated wire format (internal) representation.

Currently fetches all tables for a given namespace and filters the
results for the desired table in the client - this should be replaced
with a more efficient table-orientated schema lookup API once it is
available in IOx:

	https://github.com/influxdata/influxdb_iox/issues/4997